requirements.txt
Pillow==7.1.1
python-dotenv==0.12.0
requests==2.23.0
